scripts/requirements.txt
firebase-admin==6.6.0
python-dotenv==1.0.1
pandas==2.2.3
